refactor: 종료 상태 추상화
